cleaned up omim_scrape and added documentations
