+++ src/constants.py
@@ -1,0 +1,9 @@
+"""
+Stores all used API keys and emails
+"""
+
+APIKEY = ''
+LANGCHAIN_API = ''
+EMAI = ''
+TOGETHER_API = ""
+OPEN_ROUTER = ""
